+++ weblate/trans/tests/test_search.py
@@ -200,14 +200,20 @@ class ReplaceTest(ViewTestCase):
         self.edit_unit("Hello, world!\n", "Nazdar svete!\n")
         self.unit = self.get_unit()
 
-    def do_replace_test(self, url, confirm=True):
+    def do_replace_test(self, url, confirm=True, query=None):
+        query = query or ""
         response = self.client.post(
-            url, {"search": "Nazdar", "replacement": "Ahoj"}, follow=True
+            url, {"q": query, "search": "Nazdar", "replacement": "Ahoj"}, follow=True
         )
         self.assertContains(
             response, "Please review and confirm the search and replace results."
         )
-        payload = {"search": "Nazdar", "replacement": "Ahoj", "confirm": "1"}
+        payload = {
+            "q": query,
+            "search": "Nazdar",
+            "replacement": "Ahoj",
+            "confirm": "1",
+        }
         if confirm:
             payload["units"] = self.unit.pk
         response = self.client.post(url, payload, follow=True)
@@ -234,6 +240,21 @@ class ReplaceTest(ViewTestCase):
         )
         unit = self.get_unit()
         self.assertEqual(unit.target, "Nazdar svete!\n")
+
+    def test_replace_translated(self):
+        self.do_replace_test(
+            reverse("replace", kwargs=self.kw_translation), "is:translated"
+        )
+
+    def test_replace_nontranslated(self):
+        response = self.client.post(
+            reverse("replace", kwargs=self.kw_translation),
+            {"q": "NOT is:translated", "search": "Nazdar", "replacement": "Ahoj"},
+            follow=True,
+        )
+        self.assertNotContains(
+            response, "Please review and confirm the search and replace results."
+        )
 
     def test_replace(self):
         self.do_replace_test(reverse("replace", kwargs=self.kw_translation))